feat: Create requests.js
Modularization APIs
